Feat/policies

See merge request etu-forms/backend!1

+++ src/core/controllers/base.controller.ts
@@ -10,6 +10,9 @@ import { FindContext } from "@/core/repositories/repository.interface.js";
 import { z } from "zod";
 import { PaginatedResponse } from "@/shared/http/paginated-response.js";
 import { BasePagination } from "@/core/repositories/base.pagination.js";
+import { ForbiddenError } from "@/shared/errors/forbidden.error.js";
+import { UnauthorizedError } from "@/shared/errors/unauthorized.error.js";
+import { IResourcePolicy } from "@/core/policies/policy.interface.js";
 
 export abstract class BaseController<
   TEntity extends QueryResultRow,
@@ -19,6 +22,7 @@ export abstract class BaseController<
 > {
   protected constructor(
     protected readonly service: BaseService<TEntity, TCreate, TUpdate>,
+    protected readonly policy?: IResourcePolicy<TEntity>,
     protected readonly mapper?: IMapper<TEntity, TResponse>,
   ) {}
 
@@ -64,7 +68,14 @@ export abstract class BaseController<
       throw new NotFoundError("Entity not found");
     }
 
+    if (this.policy) {
+      if (!req.user) throw new UnauthorizedError();
+      const allowed = await this.policy.view(req.user.id, entity);
+      if (!allowed) throw new ForbiddenError();
+    }
+
     const data = this.mapper ? this.mapper.toResponse(entity) : entity;
+
     res.status(200).json(data);
   };
 
@@ -76,16 +87,40 @@ export abstract class BaseController<
 
   update = async (req: Request, res: Response): Promise<void> => {
     const id = getRouteParam(req, "id");
-    const entity = await this.service.update(id, req.body);
-    const data = this.mapper ? this.mapper.toResponse(entity) : entity;
+    const entity = await this.service.findById(id);
+
+    if (!entity) {
+      throw new NotFoundError("Entity not found");
+    }
+
+    if (this.policy) {
+      if (!req.user) throw new UnauthorizedError();
+      const allowed = await this.policy.update(req.user.id, entity);
+      if (!allowed) throw new ForbiddenError();
+    }
+
+    const updatedEntity = await this.service.update(id, req.body);
+    const data = this.mapper
+      ? this.mapper.toResponse(updatedEntity)
+      : updatedEntity;
     res.status(200).json(data);
   };
 
   delete = async (req: Request, res: Response): Promise<void> => {
     const id = getRouteParam(req, "id");
+    const entity = await this.service.findById(id);
+
+    if (!entity) {
+      throw new NotFoundError("Entity not found");
+    }
+
+    if (this.policy) {
+      if (!req.user) throw new UnauthorizedError();
+      const allowed = await this.policy.delete(req.user.id, entity);
+      if (!allowed) throw new ForbiddenError();
+    }
 
     await this.service.delete(id);
-
     res.status(204).send();
   };
 }
